requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
openai==1.58.1
pdfplumber==0.11.5
python-docx==1.1.2
numpy>=2.2.0
